uintX_t are now ints and players can move

+++ CommonNetwork/IPAddress.cpp
@@ -4,13 +4,13 @@ IPAddress::IPAddress()
 {
 }
 
-IPAddress::IPAddress(uint8_t a, uint8_t b, uint8_t c, uint8_t d, uint16_t port)
+IPAddress::IPAddress(int a, int b, int c, int d, int port)
 {
 	address = (a << 24) + (b << 16) + (c << 8) + d;
 	this->port = port;
 }
 
-IPAddress::IPAddress(uint32_t address, uint16_t port)
+IPAddress::IPAddress(int address, int port)
 {
 	this->address = address;
 	this->port = port;
