+++ CommonNetwork/IPAddress.h
@@ -7,31 +7,31 @@ class IPAddress
 {
 public:
 	IPAddress();
-	IPAddress(uint8_t a, uint8_t b, uint8_t c, uint8_t d, uint16_t port);
-	IPAddress(uint32_t address, uint16_t port);
+	IPAddress(int a, int b, int c, int d, int port);
+	IPAddress(int address, int port);
 	~IPAddress();
 
-	void SetAddress(uint8_t a, uint8_t b, uint8_t c, uint8_t d);
-	void SetAddress(uint32_t address);
-	void SetA(uint8_t a);
-	void SetB(uint8_t b);
-	void SetC(uint8_t c);
-	void SetD(uint8_t d);
-	void SetIPPort(uint16_t port);
+	void SetAddress(int a, int b, int c, int d);
+	void SetAddress(int address);
+	void SetA(int a);
+	void SetB(int b);
+	void SetC(int c);
+	void SetD(int d);
+	void SetIPPort(int port);
 
-	uint32_t GetAddress() const;
-	uint8_t GetA() const;
-	uint8_t GetB() const;
-	uint8_t GetC() const;
-	uint8_t GetD() const;
-	uint16_t GetPort() const;
+	int GetAddress() const;
+	int GetA() const;
+	int GetB() const;
+	int GetC() const;
+	int GetD() const;
+	int GetPort() const;
 	std::string ToString() const;
 
 	int operator==(IPAddress rhs);
 
 private:
-	uint32_t address;
-	uint16_t port;
+	int address;
+	int port;
 };
 
 #include "IPAddress.inl"
